rename lobby to Lobby
